sdm "conversion" on cbe. lookall working.

+++ queue_rtpipe.py
@@ -4,7 +4,7 @@
 # each job is independent but shares file system. one worker per node.
 
 from rq import Queue, Connection
-import os, argparse, time, shutil, subprocess, glob
+import os, argparse, time, shutil
 import sdmreader
 import queue_funcs as qf
 
@@ -80,38 +80,6 @@ if __name__ == '__main__':
             q = Queue(qpriority, async=False)    # ultimately need this to be on queue and depende_on search
             plotjob = q.enqueue_call(func=qf.plot_pulsar, args=(workdir, fileroot, scans), timeout=24*3600, result_ttl=24*3600)    # default TARGET intent
 
-        elif mode == 'looktest':
-            # this mode looks for file that includes filename, checks that it is completed sdm, then runs 'all'
-
-            q = Queue(qpriority, async=False)  # this will block
-            subname = os.path.split(filename)[1]
-            filejob = q.enqueue_call(func=qf.lookforfile, args=(sdmdir, subname), timeout=24*3600, result_ttl=24*3600)
-            filename = filejob.result
-            sdmjob = q.enqueue_call(func=qf.waitforsdm, args=(filename,), timeout=24*3600, result_ttl=24*3600)
-
-            # make symlink to working area
-            newfileloc = os.path.join(workdir, os.path.split(filename)[1])
-#            os.symlink(filename, newfileloc)
-            shutil.copytree(filename, newfileloc)  # copy file in
-            calscanlist = qf.getscans(newfileloc, intent='CALI')   # get calibration scans
-            calscans = ','.join([str(sc) for sc in calscanlist])   # put into CASA-like selection string
-            shutil.copyfile(newfileloc + '/Main.xml', newfileloc + '/Main_orig.xml')   # put new Main.xml in
-            subprocess.call(['choose_SDM_scans.pl', newfileloc, os.path.join(workdir, 'Main.xml'), calscans])  #  modify Main.xml
-# not copying?
-            shutil.move(os.path.join(workdir, 'Main.xml'), newfileloc + '/Main.xml')   # put new Main.xml in
-            os.makedirs(os.path.join(newfileloc, 'ASDMBinary'))
-            sc,sr = sdmreader.read_metadata(newfileloc)
-            for calscan in calscanlist:
-                print 'Copying bdf in for calscan %d' % calscan
-                bdfnum = sc[calscan]['bdfstr'].split('/')[-1]
-                bdffile = glob.glob(os.path.join(bdfdir, bdfnum))[0]
-                shutil.copyfile(bdffile, os.path.join(newfileloc, 'ASDMBinary', os.path.split(bdffile)[1]))
-            filename = newfileloc
-
-            # start jobs
-            q = Queue(qpriority)
-            caljob = q.enqueue_call(func=qf.calibrate, args=(filename, fileroot), timeout=24*3600, result_ttl=24*3600)   # can be set to enqueue when data arrives
-
         elif mode == 'lookall':
             # this mode looks for file that includes filename, checks that it is completed sdm, then runs 'all'
 
@@ -121,33 +89,23 @@ if __name__ == '__main__':
             filename = filejob.result
             sdmjob = q.enqueue_call(func=qf.waitforsdm, args=(filename,), timeout=24*3600, result_ttl=24*3600)
 
-            # make symlink to working area
+            # copy to working area and make cal-able
             newfileloc = os.path.join(workdir, os.path.split(filename)[1])
-#            os.symlink(filename, newfileloc)
-            shutil.copytree(filename, newfileloc)  # copy file in
-            calscanlist = qf.getscans(newfileloc, intent='CALI')   # get calibration scans
-            calscans = ','.join([str(sc) for sc in calscanlist])   # put into CASA-like selection string
-            shutil.copyfile(newfileloc + '/Main.xml', newfileloc + '/Main_orig.xml')   # put new Main.xml in
-            subprocess.call(['choose_SDM_scans.pl', newfileloc, os.path.join(workdir, 'Main.xml'), calscans])  #  modify Main.xml
-# not copying?
-            shutil.move(os.path.join(workdir, 'Main.xml'), newfileloc + '/Main.xml')   # put new Main.xml in
-            os.makedirs(os.path.join(newfileloc, 'ASDMBinary'))
-            sc,sr = sdmreader.read_metadata(newfileloc)
-            for calscan in calscanlist:
-                print 'Copying bdf in for calscan %d' % calscan
-                bdfnum = sc[calscan]['bdfstr'].split('/')[-1]
-                bdffile = glob.glob(os.path.join(bdfdir, bdfnum))[0]
-                shutil.copyfile(bdffile, os.path.join(newfileloc, 'ASDMBinary'))
+            if not os.path.exists(newfileloc):
+                shutil.copytree(filename, newfileloc)  # copy file in
+            else:
+                print 'File %s already in %s. Using that one...' % (newfileloc, workdir)
             filename = newfileloc
-
-            # start jobs
-            q = Queue(qpriority)
+            sdmcaljob = q.enqueue_call(func=qf.sdmascal, args=(filename,), timeout=24*3600, result_ttl=24*3600)   # make emtpy sdm workable as cal sdm
             caljob = q.enqueue_call(func=qf.calibrate, args=(filename, fileroot), timeout=24*3600, result_ttl=24*3600)   # can be set to enqueue when data arrives
+            sdmorigjob = q.enqueue_call(func=qf.sdmasorig, args=(filename,), timeout=24*3600, result_ttl=24*3600, depends_on=caljob)   # make emtpy sdm workable to search
 
-            shutil.move(filename + '/Main_orig.xml', filename + '/Main.xml')   # put orig Main.xml in
+            # add a step to fill bdfpkls?
+
+            # start non-blocking part
+            q = Queue(qpriority)
             scans = qf.getscans(filename, sources=sources, scans=scans, intent='TARGET')  # default cleans up target scans
-#            lastsearchjob = qf.search(q.name, filename, paramfile, fileroot, scans=scans, redishost=redishost, depends_on=caljob)
-            lastsearchjob = qf.search(q.name, filename, paramfile, fileroot, scans=scans, redishost=redishost)
+            lastsearchjob = qf.search(q.name, filename, paramfile, fileroot, scans=scans, redishost=redishost, depends_on=sdmorigjob)
             cleanjob = q.enqueue_call(func=qf.cleanup, args=(workdir, fileroot, scans), timeout=24*3600, result_ttl=24*3600, depends_on=lastsearchjob)  # enqueued when joblist finishes
             plotjob = q.enqueue_call(func=qf.plot_summary, args=(workdir, fileroot, scans), timeout=24*3600, result_ttl=24*3600, depends_on=cleanjob)   # enqueued when cleanup finished
